requirements.txt
mpi4py
paho-mqtt
python-dotenv
flask